// app/types/user.ts
export interface User {
    name: string;
    phonecode: string;
    phone: string;
    mail: string;
    createddate: string;
    updatedate: string;
    lixhium_pro: string;
    inappnotification: boolean;
    pushnotification: boolean;
    sendmail: boolean;
    analytics: boolean;
    phoneconfirmation: boolean;
    mailconfirmation: boolean;
    vehiclemodels: any[];
    billingAddres: any[];
} 